+++ test/hydrogen/module/core_test.clj
@@ -22,12 +22,16 @@
 (def watch-dirs
   ["some-dir" "some-other-dir"])
 
+(def reload-clj-files
+  #{:cljc})
+
 (def base-development-config
   {:duct.profile/base {:duct.core/project-ns 'foo.bar
                        :duct.core/environment :development}
    :hydrogen.module/core {:figwheel-main {:port port
                                           :host host
-                                          :watch-dirs watch-dirs}}})
+                                          :watch-dirs watch-dirs
+                                          :reload-clj-files reload-clj-files}}})
 
 (deftest module-test
   (testing "blank production config"
@@ -66,7 +70,8 @@
                                                  :open-url false
                                                  :ring-server-options {:port port :host host}
                                                  :css-dirs ["target/resources/foo/bar/public/css"]
-                                                 :watch-dirs watch-dirs}}}
+                                                 :watch-dirs watch-dirs
+                                                 :reload-clj-files reload-clj-files}}}
            (core/build-config base-development-config))))
 
   (testing "custom externs"